Back/requirements.txt
amqp==5.0.6
APScheduler==3.6.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.9
django-appconf==1.0.5
django-celery-beat==2.2.1
django-celery-email==3.0.0
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-prometheus==2.1.0
django-templated-mail==1.1.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flower==1.0.0
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
humanize==3.12.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.2
librabbitmq==2.0.0
lxml==4.6.4
MarkupSafe==2.0.1
msgpack==1.0.2
oauthlib==3.1.1
prometheus-client==0.12.0
prompt-toolkit==3.0.22
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
zope.event==4.5.0
zope.interface==5.4.0
